[librptexture] SegaPVRPrivate::loadPvrImage(): Remove the (int) cast in the SVR Index4 case.
ImageDecoder::fromLinearCI4() takes size_t for the palette buffer size,
so there's no need to cast it to int.

Fixes three clang-tidy warnings:

src/librptexture/fileformat/SegaPVR.cpp:744:60: warning: C-style casts are discouraged; use static_cast [google-readability-casting]
  744 |                                 reinterpret_cast<const uint16_t*>(svr_pal_buf.data()), (int)svr_pal_buf.size());
      |                                                                                        ^~~~~
      |                                                                                        static_cast<int>(      )
src/librptexture/fileformat/SegaPVR.cpp:773:25: warning: C-style casts are discouraged; use static_cast [google-readability-casting]
  773 |                                 svr_pal_buf.data(), (int)svr_pal_buf.size());
      |                                                     ^~~~~
      |                                                     static_cast<int>(      )
src/librptexture/fileformat/SegaPVR.cpp:823:25: warning: C-style casts are discouraged; use static_cast [google-readability-casting]
  823 |                                 svr_pal_buf.data(), (int)svr_pal_buf.size());
      |                                                     ^~~~~
      |                                                     static_cast<int>(      )

+++ src/librptexture/fileformat/SegaPVR.cpp
@@ -741,7 +741,7 @@ rp_image_const_ptr SegaPVRPrivate::loadPvrImage(void)
 				true, true,
 				pvrHeader.width, pvrHeader.height,
 				buf.get(), expected_size,
-				reinterpret_cast<const uint16_t*>(svr_pal_buf.data()), (int)svr_pal_buf.size());
+				reinterpret_cast<const uint16_t*>(svr_pal_buf.data()), svr_pal_buf.size());
 			break;
 		}
 
@@ -770,7 +770,7 @@ rp_image_const_ptr SegaPVRPrivate::loadPvrImage(void)
 			img = ImageDecoder::fromLinearCI4(px_format, false,
 				pvrHeader.width, pvrHeader.height,
 				buf.get(), expected_size,
-				svr_pal_buf.data(), (int)svr_pal_buf.size());
+				svr_pal_buf.data(), svr_pal_buf.size());
 
 			// Puyo Tools: Minimum swizzle size for 4-bit is 128x128.
 			if (pvrHeader.width >= 128 && pvrHeader.height >= 128) {
@@ -820,7 +820,7 @@ rp_image_const_ptr SegaPVRPrivate::loadPvrImage(void)
 			img = ImageDecoder::fromLinearCI8(px_format,
 				pvrHeader.width, pvrHeader.height,
 				buf.get(), expected_size,
-				svr_pal_buf.data(), (int)svr_pal_buf.size());
+				svr_pal_buf.data(), svr_pal_buf.size());
 
 			// Puyo Tools: Minimum swizzle size for 8-bit is 128x64.
 			if (pvrHeader.width >= 128 && pvrHeader.height >= 64) {
